embed::counter::count now returns uint64_t
For counters that are limited to only 32-bits, they can simply return
32-bits. For larger counters, 64-bit gives enough space.

64-bit ints can also be used for 32-bit counters with interrupt driven
overflow counters, which can be combined with the 32-bit counter in
order to get a full 64-bit count value.

+++ include/libembeddedhal/counter.hpp
@@ -20,6 +20,6 @@ public:
   virtual std::chrono::nanoseconds period() = 0;
   virtual void control(controls p_control) = 0;
   virtual bool is_active() = 0;
-  virtual uint32_t count() = 0;
+  virtual uint64_t count() = 0;
 };
 } // namespace embed
